refactored code separated advertisement create and update logic
