+++ src/main/java/ua/everybuy/buisnesslogic/config/AmazonS3ClientConfig.java
@@ -10,7 +10,6 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class AmazonS3ClientConfig {
-
     @Value("${aws.access.key}")
     private String accessKey;
