+++ src/main/java/ua/everybuy/buisnesslogic/config/RestTemplateConfig.java
@@ -4,12 +4,10 @@ import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
-
 import java.time.Duration;
 
 @Configuration
 public class RestTemplateConfig {
-
     @Bean
     public RestTemplate restTemplate(RestTemplateBuilder builder) {
         return builder
