+++ src/main/java/ua/everybuy/buisnesslogic/service/integration/UserProfileService.java
@@ -1,6 +1,5 @@
 package ua.everybuy.buisnesslogic.service.integration;
 
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
