+++ src/main/java/ua/everybuy/buisnesslogic/service/location/CityService.java
@@ -1,17 +1,15 @@
-package ua.everybuy.buisnesslogic.service;
+package ua.everybuy.buisnesslogic.service.location;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ua.everybuy.database.entity.City;
 import ua.everybuy.database.repository.CityRepository;
-
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class CityService {
-
     private final CityRepository cityRepository;
 
     public List<City> getAllCities (){
