+++ src/main/java/ua/everybuy/buisnesslogic/service/location/RegionService.java
@@ -1,17 +1,15 @@
-package ua.everybuy.buisnesslogic.service;
+package ua.everybuy.buisnesslogic.service.location;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ua.everybuy.database.entity.Region;
 import ua.everybuy.database.repository.RegionRepository;
-
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class RegionService {
-
     private final RegionRepository regionRepository;
 
     public List<Region> getAllRegions() {
