+++ src/main/java/ua/everybuy/buisnesslogic/service/photo/AmazonS3Service.java
@@ -1,4 +1,4 @@
-package ua.everybuy.buisnesslogic.service;
+package ua.everybuy.buisnesslogic.service.photo;
 
 import com.amazonaws.AmazonServiceException;
 import com.amazonaws.SdkClientException;
@@ -9,7 +9,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import ua.everybuy.database.entity.AdvertisementPhoto;
-
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.List;
@@ -18,7 +17,6 @@ import java.util.UUID;
 @Service
 @RequiredArgsConstructor
 public class AmazonS3Service {
-
     private final AmazonS3 s3Client;
 
     @Value("${aws.photo.url}")
