+++ src/main/java/ua/everybuy/routing/controller/TopLevelSubCategoryController.java
@@ -3,7 +3,7 @@ package ua.everybuy.routing.controller;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-import ua.everybuy.buisnesslogic.service.TopLevelSubCategoryService;
+import ua.everybuy.buisnesslogic.service.category.TopLevelSubCategoryService;
 import ua.everybuy.routing.dto.SubCategoryDto;
 import java.util.List;
