+++ src/main/java/ua/everybuy/routing/controller/WakeUpController.java
@@ -1,6 +1,5 @@
 package ua.everybuy.routing.controller;
 
-
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
